+++ CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added token-based pagination to `GET /orders`, `GET /products`,
   `GET /orders/{order_id}/statuses`, and `POST /products/{product_id}/opportunities`.
-- Optional and Extension STAPI Status Codes "scheduled", "held", "processing",
-  "reserved", "tasked", and "user_cancelled"
+- Optional and Extension STAPI Status Codes "scheduled", "held", "processing", "reserved", "tasked",
+  and "user_canceled"
 - Asynchronous opportunity search. If the root router supports asynchronous opportunity
   search, all products must support it. If asynchronous opportunity search is
   supported, `POST` requests to the `/products/{productId}/opportunities` endpoint will